simulator/requirements.txt
paho-mqtt
web3
python-dotenv
